refactor: remove StepsPreference
As it was unused

kept:

* convertFromJSON => convertFromStepsJSONArray
* convertToJSON => convertStepsToJSONArray

As these are used in FilteredDeckOptions
